refactor(checker): split lazy resolution fuel state
