+++ crates/tsz-checker/src/state/type_environment/mod.rs
@@ -5,4 +5,5 @@ mod application;
 mod core;
 mod formatting;
 pub(crate) mod lazy;
+mod lazy_fuel;
 mod type_node_resolution;
